+++ view/packages/wikibase-data-values-value-view/tests/src/experts/GlobeCoordinateInput.tests.js
@@ -2,7 +2,7 @@
  * @licence GNU GPL v2+
  * @author H. Snater < mediawiki@snater.com >
  */
-( function( QUnit, valueview ) {
+( function( QUnit, valueview, $ ) {
 	'use strict';
 
 	var testExpert = valueview.tests.testExpert;
@@ -13,4 +13,25 @@
 		expertConstructor: valueview.experts.GlobeCoordinateInput
 	} );
 
-}( QUnit, jQuery.valueview ) );
+	QUnit.test( 'don\'t crash with null precision', function( assert ) {
+		assert.expect( 1 );
+		var $div = $( '<div/>' ).appendTo( 'body' );
+		var expert = new valueview.experts.GlobeCoordinateInput(
+			$div,
+			new valueview.tests.MockViewState( {
+				value: { getValue: function() { return { getPrecision: function() { return null; } }; } },
+				getTextValue: 'value'
+			} )
+		);
+		expert.init();
+		expert.draw();
+		expert.focus();
+		QUnit.stop();
+		window.setTimeout( function() {
+			QUnit.start();
+			assert.ok( true );
+			$div.remove();
+		}, 300 );
+	} );
+
+}( QUnit, jQuery.valueview, jQuery ) );